refactor: extract JWT to standalone package to prevent config.Get() from affecting tests
