+++ backend/pkg/util/jwt/jwt.go
@@ -1,7 +1,7 @@
 // Copyright 2024 CloudDetail
 // SPDX-License-Identifier: Apache-2.0
 
-package util
+package jwt
 
 import (
 	"errors"
